requirements.txt
transformers
torch
python-dotenv
gitpython